+++ packages/plugin-e2e/src/auth/auth.setup.ts
@@ -1,5 +1,8 @@
 import { test as setup } from '../api';
 
-setup('authenticate', async ({ login }) => {
+setup('authenticate', async ({ login, createUser, user }) => {
+  if (user) {
+    await createUser();
+  }
   await login();
 });