+++ packages/plugin-e2e/src/models/DashboardPage.ts
@@ -1,9 +1,10 @@
-const gte = require('semver/functions/gte');
+import * as semver from 'semver';
 import { DashboardPageArgs, NavigateOptions, PluginTestCtx } from '../types';
 import { DataSourcePicker } from './DataSourcePicker';
 import { GrafanaPage } from './GrafanaPage';
 import { PanelEditPage } from './PanelEditPage';
 import { TimeRange } from './TimeRange';
+import { Panel } from './Panel';
 
 export class DashboardPage extends GrafanaPage {
   dataSourcePicker: any;
@@ -44,11 +45,42 @@ export class DashboardPage extends GrafanaPage {
   }
 
   /**
+   * Returns a Panel object for the panel with the given title. Only works for panels that currently are in the viewport.
+   *
+   * Note that this won't navigate to the panel edit page, it will only return the Panel object, which
+   * points to the locator for the panel in the dashboard page. Can be used to assert on the panel data, eg.
+   * const panel = await dashboardPage.getPanelByTitle('Table panel');
+   * await expect(panel.fieldNames).toContainText(['time', 'temperature']);
+   */
+  getPanelByTitle(title: string): Panel {
+    let locator = this.getByTestIdOrAriaLabel(this.ctx.selectors.components.Panels.Panel.title(title), {
+      startsWith: true,
+    });
+    // in older versions, the panel selector is added to a child element, so we need to go up two levels to get the wrapper
+    if (semver.lt(this.ctx.grafanaVersion, '9.5.0')) {
+      locator = locator.locator('..').locator('..');
+    }
+    return new Panel(this.ctx, locator);
+  }
+
+  /**
+   * Returns a Panel object for the panel with the given id. Only works for panels that currently are in the viewport.
+   *
+   * Note that this won't navigate to the panel edit page, it will only return the Panel object, which
+   * points to the locator for the panel in the dashboard page. Can be used to assert on the panel data, eg.
+   * const panel = await dashboardPage.getPanelByTitle('2');
+   * await expect(panel.fieldNames()).toContainText(['time', 'temperature']);
+   */
+  getPanelById(panelId: string): Panel {
+    return new Panel(this.ctx, this.ctx.page.locator(`[data-panelid="${panelId}"]`));
+  }
+
+  /**
    * Clicks the buttons to add a new panel and returns the panel edit page for the new panel
    */
   async addPanel(): Promise<PanelEditPage> {
     const { components, pages } = this.ctx.selectors;
-    if (gte(this.ctx.grafanaVersion, '10.0.0')) {
+    if (semver.gte(this.ctx.grafanaVersion, '10.0.0')) {
       await this.getByTestIdOrAriaLabel(
         components.PageToolbar.itemButton(components.PageToolbar.itemButtonTitle)
       ).click();