+++ packages/plugin-e2e/tests/as-admin-user/datasource/mocks/resource.ts
@@ -1,0 +1,9 @@
+export const GOOGLE_SHEETS_SPREADSHEETS = {
+  spreadsheets: {
+    sheet1: 'Datasource test spreadsheet',
+    sheet2: 'Google Sheets Datasource - Average Temperature',
+  },
+};
+
+export const REDSHIFT_SCHEMAS = ['public', 'test'];
+export const REDSHIFT_TABLES = ['custom-namespace1', 'custom-namespace2'];