+++ CHANGELOG.md
@@ -1,3 +1,107 @@
+# (Fri Feb 16 2024)
+
+#### 🚀 Enhancement
+
+- `@grafana/plugin-e2e@0.15.0`
+  - Plugin E2E: Improved support for multiple users and roles [#726](https://github.com/grafana/plugin-tools/pull/726) ([@sunker](https://github.com/sunker))
+
+#### Authors: 1
+
+- Erik Sundell ([@sunker](https://github.com/sunker))
+
+---
+
+# (Fri Feb 16 2024)
+
+#### 🚀 Enhancement
+
+- `@grafana/plugin-e2e@0.14.0`
+  - Plugin E2E: Add fixture that reads provisioned data source files [#740](https://github.com/grafana/plugin-tools/pull/740) ([@sunker](https://github.com/sunker))
+
+#### Authors: 1
+
+- Erik Sundell ([@sunker](https://github.com/sunker))
+
+---
+
+# (Thu Feb 15 2024)
+
+#### 🐛 Bug Fix
+
+- `@grafana/create-plugin@4.0.1`
+  - Create plugin: Fix update command [#746](https://github.com/grafana/plugin-tools/pull/746) ([@leventebalogh](https://github.com/leventebalogh))
+
+#### Authors: 1
+
+- Levente Balogh ([@leventebalogh](https://github.com/leventebalogh))
+
+---
+
+# (Wed Feb 14 2024)
+
+#### 💥 Breaking Change
+
+- `@grafana/create-plugin@4.0.0`
+  - Create-plugin: Change how the `update` command works [#707](https://github.com/grafana/plugin-tools/pull/707) ([@leventebalogh](https://github.com/leventebalogh))
+
+#### Authors: 1
+
+- Levente Balogh ([@leventebalogh](https://github.com/leventebalogh))
+
+---
+
+# (Mon Feb 12 2024)
+
+#### 🐛 Bug Fix
+
+- `@grafana/plugin-e2e@0.13.2`
+  - Plugin E2E: Fix typing in config page [#741](https://github.com/grafana/plugin-tools/pull/741) ([@sunker](https://github.com/sunker))
+
+#### Authors: 1
+
+- Erik Sundell ([@sunker](https://github.com/sunker))
+
+---
+
+# (Mon Feb 12 2024)
+
+#### 🐛 Bug Fix
+
+- `@grafana/plugin-e2e@0.13.1`
+  - Plugin E2E:: Add back health check mock [#732](https://github.com/grafana/plugin-tools/pull/732) ([@sunker](https://github.com/sunker))
+
+#### Authors: 1
+
+- Erik Sundell ([@sunker](https://github.com/sunker))
+
+---
+
+# (Sat Feb 10 2024)
+
+#### 🚀 Enhancement
+
+- `@grafana/plugin-e2e@0.13.0`
+  - Plugin E2E: Enable assertion on panel data [#688](https://github.com/grafana/plugin-tools/pull/688) ([@sunker](https://github.com/sunker))
+
+#### Authors: 1
+
+- Erik Sundell ([@sunker](https://github.com/sunker))
+
+---
+
+# (Fri Feb 09 2024)
+
+#### 🐛 Bug Fix
+
+- `@grafana/plugin-e2e@0.12.2`
+  - Plugin E2E: Generate dts files [#733](https://github.com/grafana/plugin-tools/pull/733) ([@sunker](https://github.com/sunker))
+
+#### Authors: 1
+
+- Erik Sundell ([@sunker](https://github.com/sunker))
+
+---
+
 # (Fri Feb 09 2024)
 
 #### 🐛 Bug Fix