+++ packages/create-plugin/CHANGELOG.md
@@ -1,3 +1,27 @@
+# v4.0.1 (Thu Feb 15 2024)
+
+#### 🐛 Bug Fix
+
+- Create plugin: Fix update command [#746](https://github.com/grafana/plugin-tools/pull/746) ([@leventebalogh](https://github.com/leventebalogh))
+
+#### Authors: 1
+
+- Levente Balogh ([@leventebalogh](https://github.com/leventebalogh))
+
+---
+
+# v4.0.0 (Wed Feb 14 2024)
+
+#### 💥 Breaking Change
+
+- Create-plugin: Change how the `update` command works [#707](https://github.com/grafana/plugin-tools/pull/707) ([@leventebalogh](https://github.com/leventebalogh))
+
+#### Authors: 1
+
+- Levente Balogh ([@leventebalogh](https://github.com/leventebalogh))
+
+---
+
 # v3.6.0 (Thu Feb 08 2024)
 
 #### 🚀 Enhancement